ACTIN-33: Show CNs with 1 place after decimal in ORANGE
Also standardize decimal and percentage formatting in ORANGE.

+++ orange-datamodel/src/main/java/com/hartwig/hmftools/datamodel/purple/PurpleGainLoss.java
@@ -27,7 +27,7 @@ public interface PurpleGainLoss {
 
     boolean isCanonical();
 
-    long minCopies();
+    double minCopies();
 
-    long maxCopies();
+    double maxCopies();
 }
